backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
PyJWT==2.8.0
pyserial==3.5
resend==0.8.0
requests==2.31.0
numpy==1.26.4
scikit-learn==1.4.1.post1
lightgbm==4.3.0
python-multipart==0.0.9
python-dotenv==1.0.1